Fix Artifact/Enchantments, Move CHAINING to Regular
