+++ api/src/main/kotlin/us/teaminceptus/plasmaenchants/api/enchants/PEnchantments.kt
@@ -277,6 +277,30 @@ enum class PEnchantments(
             p.addPotionEffect(PotionEffect(PotionEffectType.INCREASE_DAMAGE, 40 * level, level - 1, true))
         }),
 
+    CHAINING(
+        RANGED, 3, Action(ATTACKING) { event, level ->
+            val proj = event.damager as? Projectile ?: return@Action
+            val target = event.entity as? LivingEntity ?: return@Action
+
+            if (target.health - event.finalDamage > 0) return@Action
+
+            if (r.nextDouble() < 0.4)
+                for (i in 0 until level) {
+                    val loc = target.eyeLocation
+                    loc.pitch = 0F
+                    loc.yaw = when (i) {
+                        1 -> loc.yaw + 45F
+                        2 -> loc.yaw - 45F
+                        else -> loc.yaw
+                    }
+
+                    val newProj = target.world.spawn(loc, proj.javaClass)
+                    newProj.velocity = proj.velocity.multiply(1.25).normalize()
+                    newProj.shooter = proj.shooter
+                    newProj.setBounce(proj.doesBounce())
+                }
+        }),
+
     // Attacking Enchantments - Collectors
 
     PLAYER_COLLECTOR(
@@ -857,7 +881,7 @@ enum class PEnchantments(
         get() = info.type
 
     override val conflicts
-        get() = values().filter { conflictsP.test(it) }.toList()
+        get() = entries.filter { it != this && conflictsP.test(it) }.toList()
 
     override fun accept(e: Event, level: Int) = info.action(e, level)
 
@@ -880,7 +904,7 @@ enum class PEnchantments(
         }
 
         fun matchType(name: String): EntityType? {
-            for (type in EntityType.values())
+            for (type in EntityType.entries)
                 if (type.name.equals(name, ignoreCase = true)) return type
 
             return null
